requirements-pi.txt
gpiozero
lgpio
pyserial
